fix: nav menu items translations
now works when changing language & better use a composable
